Rename tune_params to main
